+++ src/Filter/PgSql/ILike.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Data\Db\Filter\PgSql;
+
+use Yiisoft\Data\Db\Filter\Like;
+
+class ILike extends Like
+{
+    public static function getOperator(): string
+    {
+        return 'ilike';
+    }
+}