+++ src/Processor/Not.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Data\Db\Processor;
+
+use Yiisoft\Data\Db\Filter\Not as FilterNot;
+
+final class Not extends CompareProcessor
+{
+    public function getOperator(): string
+    {
+        return FilterNot::getOperator();
+    }
+}