+++ CHANGELOG.md
@@ -3,3 +3,10 @@
 ## 1.0.0 under development
 
 - Initial release.
+
+## 24.10.2021
+
+- New Filters/Processors Not, NotEquals, Between, PgSql/ILike;
+- Argument column can be string or Yiisoft\Db\Expression\ExpressionInterface;
+- Filter Exists take only one argument;
+